+++ Ws2812AudioReactiveClient/Dsp/Volume.cs
@@ -2,9 +2,9 @@ namespace Ws2812AudioReactiveClient.Dsp;
 
 public static class Volume
 {
-    public static float LevelLinear(float[] buffer)
+    public static double LevelLinear(double[] buffer)
     {
-        var energy = 0.0f;
+        var energy = 0.0;
         uint j;
         for (j = 0; j < buffer.Length; j++)
         {
@@ -12,8 +12,8 @@ public static class Volume
         }
         return energy / buffer.Length;
     }
-    public static float DbSpl(float[] buffer)
+    public static double DbSpl(double[] buffer)
     {
-        return 10.0f * (float)Math.Log10((LevelLinear (buffer)));
+        return 10.0 * (double)Math.Log10((LevelLinear (buffer)));
     }
 }
